WeCare -using-Java-J2EE and IBM-DB2
Spread awareness about the various and most harmful cancers which can
affect women. A platform
for medical practitioners and social service organizations to engage in
community awareness and
education programs. Knowledge of measures that can be taken to prevent
these life threatening
diseases can lead to the enormous decrease in the number of victims.
Early diagnosis and treatment
can help cure the disease easily.